refactored use state for username
